+++ app/src/main/java/com/theoplayer/android/ui/demo/nitflex/NitflexUI.kt
@@ -1,12 +1,25 @@
 package com.theoplayer.android.ui.demo.nitflex
 
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.sharp.SkipNext
 import androidx.compose.material.icons.sharp.Speed
 import androidx.compose.material.icons.sharp.Subtitles
-import androidx.compose.material3.*
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.ProvideTextStyle
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -17,8 +30,16 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.theoplayer.android.api.THEOplayerConfig
 import com.theoplayer.android.api.source.SourceDescription
-import com.theoplayer.android.ui.*
+import com.theoplayer.android.ui.CurrentTimeDisplay
+import com.theoplayer.android.ui.ErrorDisplay
+import com.theoplayer.android.ui.FullscreenButton
+import com.theoplayer.android.ui.LanguageMenu
+import com.theoplayer.android.ui.LoadingSpinner
+import com.theoplayer.android.ui.PlaybackRateMenu
+import com.theoplayer.android.ui.Player
+import com.theoplayer.android.ui.UIController
 import com.theoplayer.android.ui.demo.nitflex.theme.NitflexTheme
+import com.theoplayer.android.ui.rememberPlayer
 
 @Composable
 fun NitflexUI(
@@ -27,20 +48,31 @@ fun NitflexUI(
     source: SourceDescription? = null,
     title: String? = null
 ) {
+    val player = rememberPlayer(config)
+    LaunchedEffect(key1 = player, key2 = source) {
+        player.player?.source = source
+    }
+
+    NitflexUI(modifier = modifier, player = player, title = title)
+}
+
+@Composable
+fun NitflexUI(
+    modifier: Modifier = Modifier,
+    player: Player = rememberPlayer(),
+    title: String? = null
+) {
     ProvideTextStyle(value = TextStyle(color = Color.White)) {
         UIController(
             modifier = modifier,
-            config = config,
-            source = source,
+            player = player,
             centerOverlay = {
-                val state = PlayerState.current
-                if (state?.firstPlay == true) {
+                if (player.firstPlay) {
                     LoadingSpinner()
                 }
             },
             topChrome = {
-                val state = PlayerState.current
-                if (state?.firstPlay == true) {
+                if (player.firstPlay) {
                     Row(
                         modifier = Modifier.fillMaxWidth(),
                         horizontalArrangement = Arrangement.Center,
@@ -58,12 +90,11 @@ fun NitflexUI(
                 }
             },
             centerChrome = {
-                val state = PlayerState.current
                 Row(
                     modifier = Modifier.fillMaxWidth(),
                     horizontalArrangement = Arrangement.SpaceEvenly
                 ) {
-                    if (state?.firstPlay == true) {
+                    if (player.firstPlay) {
                         NitflexSeekButton(
                             seekOffset = -10,
                             iconSize = 48.dp,
@@ -74,7 +105,7 @@ fun NitflexUI(
                         iconModifier = Modifier.size(48.dp),
                         contentPadding = PaddingValues(8.dp)
                     )
-                    if (state?.firstPlay == true) {
+                    if (player.firstPlay) {
                         NitflexSeekButton(
                             seekOffset = 10,
                             iconSize = 48.dp,
@@ -84,8 +115,7 @@ fun NitflexUI(
                 }
             },
             bottomChrome = {
-                val state = PlayerState.current
-                if (state?.firstPlay == true) {
+                if (player.firstPlay) {
                     Row(
                         modifier = Modifier.fillMaxWidth(),
                         verticalAlignment = Alignment.CenterVertically
@@ -135,11 +165,10 @@ fun NitflexUI(
                 }
             },
             errorOverlay = {
-                val state = PlayerState.current
                 Box(contentAlignment = Alignment.Center) {
                     ErrorDisplay()
                     // Ensure the user can still exit fullscreen
-                    if (state?.fullscreen == true) {
+                    if (player.fullscreen) {
                         FullscreenButton(modifier = Modifier.align(Alignment.BottomEnd))
                     }
                 }